pyproject.toml
[project]
name = "pyui-automation"
version = "0.2.0"
description = "Cross-platform desktop/Qt UI automation library"
authors = [{ name = "Ravil Shakerov", email = "xellaopromaster@yandex.ru" }]
requires-python = ">=3.8"
license = { text = "MIT" }
dependencies = [
    "black>=24.8.0",
    "colour>=0.1.5",
    "comtypes>=1.2.0; sys_platform == 'win32'",
    "pywin32>=305; sys_platform == 'win32'",
    "python-xlib>=0.31; sys_platform == 'linux'",
    "pygobject>=3.42.0; sys_platform == 'linux'",
    "pyobjc-framework-Cocoa>=9.0; sys_platform == 'darwin'",
    "pyobjc-framework-Quartz>=9.0; sys_platform == 'darwin'",
    "flake8>=5.0.4",
    "isort>=5.13.2",
    "markdown>=3.7",
    "matplotlib>=3.7.5",
    "mypy>=1.14.1",
    "numpy>=1.24.4",
    "opencv-contrib-python>=4.11.0.86",
    "opencv-python>=4.11.0.86",
    "paddleocr>=2.10.0",
    "paddlepaddle-gpu>=2.6.2",
    "pillow>=10.4.0",
    "plotly>=6.2.0",
    "psutil>=7.0.0",
    "pyautogui>=0.9.54",
    "pytest>=8.3.5",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.1",
    "pytest-xdist>=3.6.1",
    "setuptools>=75.3.2",
    "jinja2>=3.1.6",
    "pytest-asyncio>=0.24.0",
    "ty>=0.0.1a14",
    "pyrefly>=0.24.2",
]
readme = "README.md"
keywords = [
    "ui automation",
    "testing",
    "gui testing",
    "automation",
    "accessibility testing",
    "visual testing",
    "performance testing",
    "ocr",
    "cross-platform"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X"
]
urls = { "Bug Tracker" = "https://github.com/DaymaNKinG990/pyui_automation/issues", "Documentation" = "https://github.com/DaymaNKinG990/pyui_automation#readme", "Source Code" = "https://github.com/DaymaNKinG990/pyui_automation" }
scripts = { "pyui-automation" = "pyui_automation.cli:main" }

[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q --cov=pyui_automation --cov-report=term-missing"

[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[tool.setuptools]
packages = ["pyui_automation"]
